+++ CHANGELOG.md
@@ -7,3 +7,10 @@
 
 * Add led_dec and sseg_dec modules
 * Fixed sawtooth counter incrementation
+
+## [ver. 0.0.3]
+
+* Fixed FSM module
+* Fixed clk_div module
+* Added testbench unit
+* Added simulation scripts